+++ _includes/header.njk
@@ -1,6 +1,6 @@
 {% set items = [
   { 'Approach': '/approach/' },
-  { 'Projects': '/projects/' },
+  { 'Research areas': '/projects/' },
   { 'Publications': '/publications/' },
   { 'People': '/people/' },
   { 'Outreach': '/outreach/academic-programs/' },
@@ -43,5 +43,5 @@
 {% if page.url == '/' %}
 <div style="position: absolute; display: block; border: 0; padding: 0; margin: 0; width: 100%; height: 0.2rem; top: 3.6rem; background-color: #f0f0f0;">&nbsp;</div>
 {% else %}
-<div style="position: absolute; display: block; border: 0; padding: 0; margin: 0; width: 100%; height: 0.2rem; top: 3.6rem; background: linear-gradient(70deg, var(--gold-9) 30%, rgba(0,0,0,0) 30%), linear-gradient(30deg, var(--indigo-8) 60%, var(--blue-8) 60%);">&nbsp;</div>
+<div style="position: absolute; display: block; border: 0; padding: 0; margin: 0; width: 100%; height: 0.2rem; top: 3.6rem; background: linear-gradient(70deg, var(--gold-8) 30%, rgba(0,0,0,0) 30%), linear-gradient(30deg, var(--orange-8) 60%, var(--cyan-8) 60%);">&nbsp;</div>
 {% endif %}
