+++ _includes/homepage_features.njk
@@ -10,9 +10,9 @@
 {% for feature in features -%}
 
 {% if loop.index % 3 == 0 %}
-  {% set border_css = 'var(--blue-7)' %}
+  {% set border_css = 'var(--cyan-7)' %}
 {% elseif loop.index % 2 == 0 %}
-  {% set border_css = 'var(--indigo-7)' %}
+  {% set border_css = 'var(--orange-7)' %}
 {% else %}
   {% set border_css = 'var(--gold-8)' %}
 {% endif %}
