+++ _includes/layouts/content.njk
@@ -2,9 +2,6 @@
 layout: skaffold
 ---
 
-  <nav class="nav" id="nav">
-{% include 'navigation.njk' %}
-  </nav>
   <article class="content" id="content">
     {{ content | latex | safe }}
   </article>
