--- _includes/layouts/skaffold_nav.njk
@@ -1,24 +1,0 @@
----
-layout: skaffold
----
-
-<nav class="nav" id="nav">
-  {% if navigation %}
-  <ul>
-    {% for item in navigation -%}
-    {% if item.path in page.url -%}
-    <li class="active"><a href="{{ item.path }}">{{ item.label }}</a></li>
-    <ul class="subnav">  
-      {% for subitem in lookup[ item.path ] %}
-      <li class="subnav {% if subitem.path in page.url %}active{% endif %}"><a href="{{ subitem.path }}">{{ subitem.label }}</a></li><br>
-      {% endfor %}
-    </ul>
-    {% else %}
-    <li><a href="{{ item.path }}">{{ item.label }}</a></li>
-    {% endif -%}
-    {% endfor -%}
-  </ul>
-  {% endif %}
-</nav>
-
-{{ content | latex | safe }}
