+++ css/components.css
@@ -141,7 +141,9 @@ div.homepage_header {
   text-align: center;
   /* background: linear-gradient(70deg, var(--indigo-9) 30%, rgba(0,0,0,0) 30%), linear-gradient(30deg, var(--indigo-8) 60%, var(--indigo-7) 60%); */
   /* background: linear-gradient(70deg, var(--gold-9) 30%, rgba(0,0,0,0) 30%), linear-gradient(30deg, var(--indigo-8) 60%, var(--blue-8) 60%); */
-  background: linear-gradient(70deg, var(--gold-9) 30%, rgba(0,0,0,0) 30%), linear-gradient(30deg, var(--indigo-8) 60%, var(--blue-8) 60%);
+  /* background: linear-gradient(70deg, var(--gold-9) 30%, rgba(0,0,0,0) 30%), linear-gradient(30deg, var(--indigo-8) 60%, var(--blue-8) 60%); */
+  background: linear-gradient(70deg, var(--gold-8) 30%, rgba(0,0,0,0) 30%), linear-gradient(30deg, var(--orange-8) 60%, var(--cyan-8) 60%);
+  
   padding: 1.5rem 3rem;
   text-shadow: 0.05rem 0.05rem white;
 }
