+++ outreach/academic-programs/interns/index.njk
@@ -1,6 +1,6 @@
 ---
 layout: skaffold
-title: Interns
+title: Research Student Internships
 ---
 
 <nav class="nav" id="nav">
@@ -8,7 +8,7 @@ title: Interns
 
 <article class="content" id="content">
 
-<h1>Internships</h1>
+<h1>Research Student Internships</h1>
 
 As part of Cloudflare’s effort to build collaborations with academia, we host research focused internships all year long. 
 <br>
