+++ outreach/academic-programs/researchers/index.njk
@@ -1,6 +1,6 @@
 ---
 layout: skaffold
-title: Visiting Researchers
+title: Visiting Researcher Program
 ---
 
 <nav class="nav" id="nav">
@@ -8,7 +8,7 @@ title: Visiting Researchers
 
 <article class="content" id="content">
 
-<h1>Visiting Researchers</h1>
+<h1>Visiting Researcher Program</h1>
 
 Cloudflare Research is open to collaborate with the global research community to help build a better Internet in multiple areas. Current research focuses on applied cryptography, privacy, network protocols and architecture, measurement and performance evaluation, and, increasingly, distributed systems. 
 <br />
