+++ outreach/events/2022-05-30-eurocrypt.md
@@ -1,0 +1,7 @@
+---
+tags: event
+title: Eurocrypt 2022
+link: https://eurocrypt.iacr.org/2022/
+end: 2022-06-03
+description: Eurocrypt 2022 is a hybrid event and will take place in Trondheim, Norway on May 30 to June 3, 2022. Eurocrypt 2022 is organized by the International Association for Cryptologic Research (IACR).
+---
