+++ outreach/events/2022-07-19-ethcc.md
@@ -1,6 +1,6 @@
 ---
 tags: event
-title: Ethereum Community Conference (EthCC)
+title: Ethereum Community Conference 5 (EthCC)
 link: https://ethcc.io/
 end: 2022-07-21
 description: The Ethereum Community Conference (EthCC) is the largest annual European Ethereum event focused on technology and community.
