+++ projects/application-privacy/index.njk
@@ -1,5 +1,5 @@
 ---
-layout: skaffold_nav
+layout: skaffold
 title: Application Privacy
 ---
 
@@ -13,6 +13,7 @@ Some text
 <br />
 <br />
 
+<h2>Projects</h2>
 <ul>
 {% for item in lookup[ page.url ] %}
   <li><a href="{{ item.path }}">{{ item.label }}</a></li>
