+++ projects/index.njk
@@ -1,6 +1,6 @@
 ---
 layout: skaffold
-title: Projects
+title: Research areas
 ---
 
 <nav class="nav" id="nav">
@@ -19,7 +19,7 @@ title: Projects
 
 <article class="content" id="content">
 
-<h1>Projects</h1>
+<h1>Research areas</h1>
 Cloudflare Research embarks upon projects that help shape the future of the Internet at large and the systems that drive it. 
 <br />
 <br />
