+++ projects/network-privacy/index.njk
@@ -1,5 +1,5 @@
 ---
-layout: skaffold_nav
+layout: skaffold
 title: Network Privacy
 ---
